add gui.h, core.h and pipeline_editor.h/.cpp and remove warning

+++ src/exceptions.h
@@ -29,10 +29,10 @@ private:
 
 class InvalidArgN : public std::logic_error {
 public:
-    InvalidArgN(const int& expect_n, const int& input_n)
+    InvalidArgN(const size_t& expect_n, const size_t& input_n)
         : std::logic_error("InvalidArgN"),
-          inputN(input_n),
-          expectedN(expect_n),
+          inputN(int(input_n)),
+          expectedN(int(expect_n)),
           err_message(std::string("StandardFunction::build() failed : Invalid "
                                   "Number of Arguments ( expect ") +
                       std::to_string(expectedN) +
